requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.8.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
html5lib==1.1
idna==3.10
lxml==6.0.1
mysqlclient==2.2.7
oscrypto==1.3.0
pillow==11.3.0
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.0.0
python-bidi==0.6.6
python-decouple==3.8
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
